feat: refactoring folder structure for incoming module
