+++ tests/core/modules/helm_management/helm_vars_management/test_deployment_vars_creator.py
@@ -6,7 +6,7 @@ from tests.core.modules.config_management.fixtures import fixture_config_setting
 from tests.core.modules.helm_management.fixtures import fixture_filtered_deployment_helm_vars_data
 
 # Domain
-from core.modules.helm_management.helm_vars_management import HelmDeploymentVarsCreator
+from core.modules.helm_management.helm_deployment_vars import HelmDeploymentVarsCreator
 
 
 def test_helm_deployment_vars_data_creator(fixture_config_settings_data, fixture_filtered_deployment_helm_vars_data):
